inference-app/rest-api-i/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
asyncpg==0.30.0
    # via rest-api (pyproject.toml)
click==8.2.1
    # via uvicorn
colorama==0.4.6
    # via click
fastapi==0.116.1
    # via rest-api (pyproject.toml)
filelock==3.19.1
    # via torch
fsspec==2025.9.0
    # via torch
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
jinja2==3.1.6
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.2
    # via torchvision
pillow==11.3.0
    # via torchvision
prometheus-client==0.22.1
    # via rest-api (pyproject.toml)
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via rest-api (pyproject.toml)
python-multipart==0.0.20
    # via rest-api (pyproject.toml)
setuptools==80.9.0
    # via torch
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via rest-api (pyproject.toml)
starlette==0.47.3
    # via
    #   rest-api (pyproject.toml)
    #   fastapi
sympy==1.14.0
    # via torch
torch==2.8.0
    # via
    #   rest-api (pyproject.toml)
    #   torchvision
torchvision==0.23.0
    # via rest-api (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via rest-api (pyproject.toml)
